third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/execution-timing/116-expected.txt

This is a testharness.js-based test.
FAIL scheduler: adding script to head of frameset document assert_array_equals: property 0, expected "document.body: <FRAMESET>" but got "document.body: <BODY>"
Harness: the test ran to completion.
